Changed db folder name + some additional changes, wont change db initialization for now.
