+++ webapp/src/components/Login.test.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { render, fireEvent, screen, waitFor, act } from '@testing-library/react';
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import Loginview from './Login';
 
 const mockAxios = new MockAdapter(axios);
